Rename 'teamsworkflow' to 'teams-workflows'
