+++ alerting/provider/teamsworkflows/teamsworkflows_test.go
@@ -1,4 +1,4 @@
-package teamsworkflow
+package teamsworkflows
 
 import (
 	"encoding/json"
